refactor event structure
